39.2.Showing Custom Children(Section 3: Communicating with Props)
So because you want to reference a javascript variable here Propst children are put in my curly braces

and say prop's not children like so and then I'm going to remove that log.

OK.

So I'll now save this and if I now flip back over we can see that my comment detail is now showing up

inside of the approval card just as we intended.

Awesome.

So that's pretty much it.

That is the strategy that we use when we want to make truly reusable components if we ever want to show

some component or some custom content inside of a component then we will take the child.

We will wrap it with a parent component and then that child component is going to show up inside of

the child on the prop's children property.

It's always going to show up on the Propst not children so that we can then take that Propst start children

property and reference it from anywhere inside of our GSX.

So if we wanted to instead show the child at say the very bottom of the card then I could cut that out

and paste Propp start children with the curly braces down here instead.

And now we'll see that the comment shows up down here at the very bottom of the card which looks kind

of funky but I think you now get the idea for us in particular.

We wanted to show it inside of the div with class name content so that it looks like that.

Where is it.

That right there.

All right so that's pretty much it.

That's how we create some reusable components.

So now we know how to do that.

I would really like it if we took a pause right here in between this video and the next you tried to

take this approval card component and use it to wrap both of our other comment details as well.

So all you do is essentially do the same thing we did right there to show an instance of comment detail

inside of our approval card.

So a quick pause and we'll go over the solution to this in the next video.